Make Package Be Location Not Sensor
Makes more sense and ties to subsystem. Future repackaging could combine them to one factory.
