+++ src/main/java/frc/robot/sensors/SensorsFactory.java
@@ -17,9 +17,9 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import frc.robot.sensors.elevator.ElevatorSensorFactory;
 import frc.robot.sensors.gyro.GyroFactory;
 import frc.robot.sensors.incrementer.IncrementerLoadedSensorFactory;
-import frc.robot.sensors.turretlocation.TurretLocationFactory;
+import frc.robot.sensors.turret.TurretLocationFactory;
 import frc.robot.sensors.vision.VisionFactory;
-import frc.robot.sensors.walldistance.WallDistanceSensorFactory;
+import frc.robot.sensors.chassis.*;
 import frc.robot.telemetry.TelemetryManager;
 import frc.robot.telemetry.TelemetryNames;
 import frc.robot.utils.PKStatus;
